% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_prep.R
\name{data_prep}
\alias{data_prep}
\title{data_prep}
\usage{
data_prep(data)
}
\arguments{
\item{data}{Data frame.}
}
\description{
Adds id and key columns (necessary for model estimation and
  decomposition) to model data frame.
}
\examples{
tidyecon:::data_prep(mtcars)
}
