% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_lm.R
\name{decomp}
\alias{decomp}
\title{decomp
Generic decomp function}
\usage{
decomp(model, ...)
}
\arguments{
\item{model}{Model object (e.g. output from \code{lm_fit})}

\item{...}{list of additional options:
\describe{
  \item{variable}{data.frame with "categeory" column}
  \item{resolve}{"actual" or "fitted". Default is "actual"}
}}
}
\value{
A data.frame.
}
\description{
decomp
Generic decomp function
}
