% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_lm.R
\name{decomp_lm}
\alias{decomp_lm}
\title{decomp_lm}
\usage{
decomp_lm(x, y, options = list())
}
\arguments{
\item{x}{model matrix}

\item{y}{target variable}

\item{options}{list of additional options:
\describe{
  \item{coefficients}{coefficient vector}
  \item{id}{observation ids}
  \item{resolve}{either \code{"actual"} or \code{"fitted"}}
}}
}
\description{
linear model decomposition
}
\examples{
x = cbind(intercept = 1, mtcars[-match("mpg", names(mtcars))])
y = mtcars[["mpg"]]
decomp = decomp_lm(x, y)
# plot(decomp)
}
