% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_lm.R
\name{decomp_options_prep}
\alias{decomp_options_prep}
\title{decomp_options_prep}
\usage{
decomp_options_prep(model, options)
}
\arguments{
\item{model}{an lm_fit object}

\item{options}{a list of additional options (resolve, )}
}
\description{
decomp_options_prep
}
