% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(data, model_spec, options, env = parent.frame())
}
\arguments{
\item{data}{data.frame}

\item{model_spec}{Model specification data frame
\describe{
  \item{expr}{expressions for explanatory variables}
  \item{lower}{(optional) coefficient lower boundaries}
  \item{upper}{(optional) coefficient upper boundaries}
  \item{category}{(optional) decomp categories}
  \item{link}{(optional) coefficient links}
}}

\item{options}{list of options
\describe{
  \item{y}{target variable expression}
  \item{w}{(optional) regression weight expression}
  \item{link}{(optional) either "linear" (default) or "binomial"}
}}
}
\description{
wrapper for model estimation using lm_fit
}
\examples{
data = mtcars
model_spec = data.frame(expr = c("1", "cyl", "hp", "disp"))
options = list(y = "mpg")
estimate(data, model_spec, options)
}
