% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.R
\name{lm_fit}
\alias{lm_fit}
\title{lm_fit}
\usage{
lm_fit(x, y, w = NULL, lower = rep(-Inf, ncol(x)), upper = rep(Inf,
  ncol(x)), tol = 1e-07)
}
\arguments{
\item{x}{design matrix (N x n)}

\item{y}{target (vector or N x 1 matrix)}

\item{w}{vector of weights}

\item{lower}{lower bounds for coefficients}

\item{upper}{upper bounds for coefficients}

\item{tol}{QR decomposition tolerance}
}
\description{
lm_fit
}
\examples{
lm_fit(as.matrix(cbind(1, mtcars[c("cyl", "disp", "hp")])), mtcars[["mpg"]])
}
