% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_id}
\alias{make_id}
\title{make_id}
\usage{
make_id(X)
}
\arguments{
\item{X}{Data frame or model matrix}
}
\description{
makes row ids for a matrix or data.frame using existing row
  names or row numbers
}
\examples{
tidyecon:::make_id(mtcars)
tidyecon:::make_id(airquality)
}
