% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_prep.R
\name{model_spec_prep}
\alias{model_spec_prep}
\title{model_spec_prep}
\usage{
model_spec_prep(model_spec)
}
\arguments{
\item{model_spec}{data frame or vector containing variable expressions.}
}
\description{
Checks model spec for required fields and infers additional
  inputs where possible. See estimate() documentation.
}
\examples{
tidyecon:::model_spec_prep(data.frame(expr = "1"))
}
