% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_prep.R
\name{options_prep}
\alias{options_prep}
\title{options_prep}
\usage{
options_prep(options)
}
\arguments{
\item{options}{List of options.}
}
\description{
Checks options list for required inputs. See estimate()
  documentation.
}
\examples{
tidyecon::options_prep(list(y = "mpg"))
}
