% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.decomp}
\alias{plot.decomp}
\title{plot-decomp}
\usage{
\method{plot}{decomp}(x, ..., base_size = 12)
}
\arguments{
\item{x}{a decomp data.frame}

\item{...}{additional unused arguments}

\item{base_size}{ggplot theme size}
}
\description{
plot-decomp
}
