% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vnums}
\alias{vnums}
\title{vnums}
\usage{
vnums(n, factor = FALSE)
}
\arguments{
\item{n}{number of labels}

\item{factor}{(optional) logical value to convert output to factor}
}
\description{
vnums
}
\examples{
tidyecon:::vnums(5)
tidyecon:::vnums(10)
tidyecon:::vnums(10, factor = TRUE)
}
