% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.decomp}
\alias{plot.decomp}
\title{plot-decomp}
\usage{
\method{plot}{decomp}(decomp, base_size = 12)
}
\arguments{
\item{decomp}{a decomp data.frame}

\item{base_size}{base size for ggplot theme}
}
\description{
plot-decomp
}
\examples{

}
